New buildservice config handling.
Expect the config object to have an obs section with an oscrc option which
specified where the oscrc to be used lives. If not None is passed to
BuildService which allows it to look in current user's home.

The section can come from the common to all participants /etc/skynet/skynet.conf
or the participant's own /etc/skynet/<pname>.conf