[update_patterns] Listen to BUILD_SUCCESS because SRs don't cause a _COMMIT
Also change the interface to take both repository and arch because the
buildservice API requires both; this also meant fixing the "pick one"
logic (in case the params are not supplied) to be a lot smarter.
It's much more reliable now though: if the package is only built in
one repository then it will pick that one.