changelog: reorder history to preserve time-space continuum
Now that we generata rpm .changes from the debian/changelog, the rpm
tools complain that some of the entries are out of order.
This is a consequence of the way we merged branches.

Instead of investigating the true history deeply, I just changed some
timestamps to make them in order again.