update_ts process to use l10n_update_ts
The process gets triggered by a BUILD_SUCCESS event from OBS.
It then waits for the repository to be published and searches
for <package-name>-ts-devel providers. If any are found, the
l10n_update_ts participant is called to fetch the new templates
and upload them to a git repo.