[robo3] robogrator shim with fanout exchange and named durable queue
'fanout' ensures messages are delivered to all subscribers in case
anyone else wants to listen to obs events.

A named queue allows durability of exchange and persistence of
messages which means they're not lost over rabbit restarts etc

Exceptions are caught and the connection is retried after a small
pause.

Messages are not ack'ed to Rabbit until they have been processed.

Runs under python2 and python3 but RuoteAMQP is still python2 only

Signed-off-by: David Greaves <david.greaves@jolla.com>